# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A list of the currently provided Interfaces to interact with FairOS decentralised file system(dfs), implementing user, pod, file system, key value store and document store",
        "title": "FairOS-dfs server",
        "contact": {
            "name": "Sabyasachi Patra",
            "email": "sabyasachi@datafund.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v0.0.0"
    },
    "paths": {
...
        "/v1/pod/delete": {
            "delete": {
                "description": "PodDeleteHandler is the api handler to delete a pod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pod"
                ],
                "summary": "Delete pod",
                "parameters": [
                    {
                        "description": "pod name and user password",
                        "name": "pod_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.PodNameRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "cookie parameter",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    }
                }
            }
        },
...
        "/v1/pod/new": {
            "post": {
                "description": "PodCreateHandler is the api handler to create a new pod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pod"
                ],
                "summary": "Create pod",
                "parameters": [
                    {
                        "description": "pod name and user password",
                        "name": "pod_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.PodNameRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "cookie parameter",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    }
                }
            }
        },
        "/v1/pod/open": {
            "post": {
                "description": "PodOpenHandler is the api handler to open pod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pod"
                ],
                "summary": "Open pod",
                "parameters": [
                    {
                        "description": "pod name and user password",
                        "name": "pod_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.PodNameRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "cookie parameter",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    }
                }
            }
        },
        "/v1/pod/open-async": {
            "post": {
                "description": "PodOpenAsyncHandler is the api handler to open pod asynchronously",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pod"
                ],
                "summary": "Open pod",
                "parameters": [
                    {
                        "description": "pod name and user password",
                        "name": "pod_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.PodNameRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "cookie parameter",
                        "name": "Cookie",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    }
                }
            }
        },
...
        "/v2/user/present": {
            "get": {
                "description": "checks if the new user is present in the new ENS based authentication",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Check if user is present",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user name",
                        "name": "userName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.PresentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.PodNameRequest": {
            "type": "object",
            "properties": {
                "podName": {
                    "type": "string"
                }
            }
        },
...
        "api.PresentResponse": {
            "type": "object",
            "properties": {
                "present": {
                    "type": "boolean"
                }
            }
        },
...
        "api.response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}